+++ Components/JustGiving/Events/JustGivingCommandSendDonations.cs
@@ -1,0 +1,10 @@
+﻿using Slipstream.Shared;
+
+namespace Slipstream.Components.JustGiving.Events
+{
+    public class JustGivingCommandSendDonations : IEvent
+    {
+        public string EventType => nameof(JustGivingCommandSendDonations);
+        public IEventEnvelope Envelope { get; set; } = new EventEnvelope();
+    }
+}