+++ Components/JustGiving/Lua/IJustGivingInstanceThread.cs
@@ -1,0 +1,8 @@
+﻿using Slipstream.Shared.Lua;
+
+namespace Slipstream.Components.JustGiving.Lua
+{
+    public interface IJustGivingInstanceThread : ILuaInstanceThread
+    {
+    }
+}