+++ CHANGELOG.md
@@ -14,6 +14,7 @@
  - Added "Clear All Events" option when right-clicking in the events tab
  - Added "Save events to file" option when right clicking in the events tab
  - Generate correct lua handler code for custom events
+ - New component: JustGiving - monitor fundraisings and get events at new donations
 
 ## [0.10.0](https://github.com/dennis/slipstream/releases/tag/v0.10.0) (2021-10-07)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.9.0...v0.10.0)